Start modularizing dccrg.
Move grid length related stuff to dccrg_length.hpp,
cell id mapping related stuff to dccrg_mapping.hpp
and periodicity stuff to dccrg_topology.hpp.
Also make the above functionality available to dccrg
by storing the classes as variables in dccrg.

Change the dccrg API:
Geometry related functions are now available through
the geometry member of dccrg. Setting the geometry
of the grid is now optional but the initialize
function must be given the length of the grid in
cells of refinement level 0.

See the doxygen documentation and the programs in
examples and tests directories for the specifics.